Crate zed_cryptopals_challenge

Zed is working through Cryptopals Challenge

Zed is working through https://cryptopals.com/ to learn how to actually use rust.

Challenges are solved by writing and running tests in relevant modules. They’re not solved or addressed directly. For example, you can’t find a crate::challenges::set1::challenge3 module that directly addresses https://cryptopals.com/sets/1/challenges/3.

That’s because the Cryptopals challenges build upon each other to form a comprehensive cryptographic toolkit.

Except: Some challenges that result in one-off code are addressed direcctly in crate::challenges.

References

https://dev.to/tiemen/implementing-base64-from-scratch-in-rust-kb1

RFC 4648

Modules

challenges

One-off implementations for specific Cryptopals challenges. Code specific to certain challenges in https://cryptopals.com/.

codec

Encoding formats implemented to provide encode and decode utilities.

crack

Implementation of various encryption and decryption tasks

utils

Shared utilities for this crate. Shared utilities