Trait zed_cryptopals_challenge::codec::adapter::CodecAPI

pub trait CodecAPI {
    // Required methods
    fn raw_encode(&self, v: &[u8]) -> Vec<u8>;
    fn raw_to_plain(&self, v: &[u8]) -> Vec<u8>;

    // Provided methods
    fn get_chunksize(&self) -> usize { ... }
    fn encode(&self, data: &[u8]) -> Vec<u8> { ... }
    fn encode_to_string(&self, data: &[u8]) -> String { ... }
    fn to_plain(&self, data: &[u8]) -> Vec<u8> { ... }
    fn to_plain_string(&self, data: &[u8]) -> String { ... }
}

Required Methods

fn raw_encode(&self, v: &[u8]) -> Vec<u8>

Provide low-level transformation of content from one encoding format to another by manipulating the bytes that represent that content, and then applying the mapping provided by CodePointMap::map_codepoint_to_plain.

fn raw_to_plain(&self, v: &[u8]) -> Vec<u8>

Provide low-level transformation of content from one encoding format to another by manipulating the bytes that represent that content, and then applying the mapping provided by CodePointMap::map_codepoint_to_plain.

Provided Methods

fn get_chunksize(&self) -> usize

Helper that just stores and returns the number of bytes this Codec instance should expect to operate on for a given encoding/decoding implementation.

Only called internally by CodecAPI::encode to provide chunks of this size for CodecAPI::raw_encode to operate over.

By default, returns 4 for the 4 byte chunks we expect to contain UTF8 characters. Change this when working with encodings that expect a different chunk size. For example:

crate::codec::b64::Base64 changes this to 3 because Base64 encoding expects 3 byte chunks

crate::codec::hex::Hexadecimal changes this to 2 because Hexadecimal encoding expects 2 byte chunks

Used in lieu of being able to declare const values in a trait.

fn encode(&self, data: &[u8]) -> Vec<u8>

Encode a byte slice using CodecAPI::raw_encode.

fn encode_to_string(&self, data: &[u8]) -> String

Convenience function that wraps CodecAPI::encode to encode a byte slice as a String in the target encoding format.

fn to_plain(&self, data: &[u8]) -> Vec<u8>

Decode a byte slice using CodecAPI::raw_to_plain.

fn to_plain_string(&self, data: &[u8]) -> String

Convenience function that wraps CodecAPI::to_plain to decode a byte slice as a String in the target encoding format.